Fix clippy warning: remove unnecessary u32 casts in transistor solver
WHY: Maintain zero-warning CI gate requirement

WHAT: Changed iteration as u32 to iteration in loop that's already u32
      Removed 2 unnecessary type casts in TransistorSimulator::drive_input

HOW: Direct replacement of 'iteration as u32 + 1' with 'iteration + 1'

+++ mcs4-emu/crates/mcs4-core/src/transistor_solver.rs
@@ -256,11 +256,11 @@ impl TransistorSimulator {
 
             if !changed {
                 converged = true;
-                self.iteration_count = iteration as u32 + 1;
+                self.iteration_count = iteration + 1;
                 break;
             }
 
-            self.iteration_count = iteration as u32 + 1;
+            self.iteration_count = iteration + 1;
         }
 
         converged
